requirements_local.txt
fastapi
langserve
sse_starlette
openai
langchain
pydantic==1.10.9
translator